fix: WeatherDetailView 구현
